requirements.txt
requests
pandas
openpyxl
numpy
tqdm
pyyaml
